Move type definitions local to their respective domains
branch:
