+++ src/mpyl/reporting/formatting/markdown.py
@@ -9,13 +9,14 @@ from junitparser import TestSuite
 
 from ...project import Stage, Project
 from ...steps import Output, ArtifactType
-from ...steps.models import DeployedHelmAppSpec, JunitTestSpec
+from ...steps.deploy.kubernetes import DeployedHelmAppSpec
 from ...steps.run import RunResult
 from ...steps.steps import StepResult
 from ...utilities.junit import (
     TestRunSummary,
     sum_suites,
     to_test_suites,
+    JunitTestSpec,
 )
