+++ src/mpyl/steps/deploy/k8s/chart.py
@@ -55,7 +55,7 @@ from .resources.traefik import (
     V1AlphaMiddleware,
     HostWrapper,
 )  # pylint: disable = no-name-in-module
-from ...models import Input, ArtifactType, DockerImageSpec
+from ...models import Input, ArtifactType
 from ....project import (
     Project,
     KeyValueProperty,
@@ -74,6 +74,7 @@ from ....project import (
     Metrics,
 )
 from ....stages.discovery import DeploySet
+from ....utilities.docker import DockerImageSpec
 
 yaml = YAML()
