+++ src/mpyl/utilities/docker/__init__.py
@@ -9,10 +9,20 @@ from traceback import print_exc
 from typing import Dict, Optional, Iterator, cast, Union, List
 from python_on_whales import docker, Image, Container, DockerException
 from python_on_whales.exceptions import NoSuchContainer
+from ruamel.yaml import yaml_object, YAML
 
 from ..logging import try_parse_ansi
 from ...project import Project
-from ...steps.models import Input
+from ...steps.models import Input, ArtifactSpec
+
+yaml = YAML()
+
+
+@yaml_object(yaml)
+@dataclass
+class DockerImageSpec(ArtifactSpec):
+    yaml_tag = "!DockerImageSpec"
+    image: str
 
 
 @dataclass(frozen=True)
