+++ tests/reporting/test_simple_reporting.py
@@ -1,5 +1,5 @@
+from src.mpyl.utilities.junit import JunitTestSpec
 from src.mpyl.reporting.formatting.text import to_string, to_test_report
-from src.mpyl.steps.models import JunitTestSpec
 from tests import root_test_path
 from tests.reporting import create_test_result
 from tests.test_resources.test_data import assert_roundtrip
@@ -14,7 +14,7 @@ class TestReporting:
         assert_roundtrip(self.test_resource_path / "simple_run.txt", simple_report)
 
     def test_should_convert_test_report_to_string(self):
-        spec = JunitTestSpec(self.test_resource_path)
+        spec = JunitTestSpec(str(self.test_resource_path))
         test_report = to_test_report(spec)
         assert_roundtrip(
             self.test_resource_path / "simple_test.txt", test_report, overwrite=False
